src/remote-control-application/requirements.txt
CoAPthon3==1.0.2
mysql-connector-python==8.2.0
protobuf==4.21.12
